+++ Basis/Packages/dev.hai-vr.basis.comms/Scripts/Systems/Runtime/ScriptableObjects/HVRAddress.cs
@@ -2,9 +2,14 @@ using UnityEngine;
 
 namespace HVR.Basis.Comms
 {
-    [CreateAssetMenu(fileName = "HVRAddress", menuName = "HVR/Address")]
+    [CreateAssetMenu(fileName = "HVRAddress", menuName = "HVR.Basis/Comms")]
     public class HVRAddress : ScriptableObject
     {
         public string path;
+
+        public string AsPath()
+        {
+            return !string.IsNullOrWhiteSpace(path) ? path : name;
+        }
     }
 }